Major restructuring in FluentCMS project
This commit represents a significant overhaul within the FluentCMS project, focusing on the reorganization of error handling mechanisms, project and solution structure adjustments, and the reconsideration of authentication context management. Key changes include the removal and subsequent re-addition of the `AppApiClientException`, `AppError`, `AppException` classes, and the `ExceptionCodes` constants file, indicating a major revision in how errors and exceptions are handled. The `FluentCMS.Common.csproj` project file was deleted, and the project was removed from the `FluentCMS.sln` solution file, suggesting a consolidation or restructuring of the project. Adjustments in project GUIDs, configurations, and references reflect an optimization of the project dependencies and architecture. The `IAuthContext` interface was also removed and then re-added, highlighting a reevaluation of the authentication context within the application. These changes collectively aim to enhance the maintainability, error handling, and overall structure of the FluentCMS project.