#1900 implement message bus using MediatR (#1914)
* #1900 implement message bus using MediatR

* Fix action name in SiteService Delete method

Updated the `Delete` method in the `SiteService` class to change the action name from `SiteUpdated` to `SiteDeleted` when publishing a message after a site is deleted.

* Refactor service registration for message bus

Refactored the `AddApplicationServices` method in `ServiceExtensions` by moving MediatR and `InMemoryMessagePublisher` registration to a new `AddInMemoryMessageBus` method in `MessageBusServiceExtensions`. This improves code organization and maintains existing functionality.