2047 final review for mongodb repository and litedb repository (#2048)
* Refactor repositories and improve type safety

- Removed unnecessary null-forgiving operator in `LayoutService`.
- Changed `DeleteMany` return type in `IEntityRepository` to non-nullable.
- Made `GetCollectionName` static in `EntityRepository`.
- Used C# 8.0 range operator in `GetAll` and `GetByIds`.
- Updated `DeleteMany` in `EntityRepository` to match interface.
- Added index creation on `Id` field for efficient querying.
- Modified insert/update to disable document validation bypass.
- Replaced `FindOneAndReplaceAsync` with `ReplaceOneAsync`.
- Added `cancellationToken` to `ToListAsync` in `PermissionRepository`.

These changes enhance code safety, performance, and readability.

* Remove BsonCollection from EntityRepository<TEntity>

The BsonCollection field of type ILiteCollectionAsync<BsonDocument>
has been removed from the EntityRepository<TEntity> class, along
with its initialization in the constructor. This change indicates
that BsonCollection is no longer needed or used within the class.

* Remove BsonCollection from EntityRepository

The BsonCollection field of type IMongoCollection<BsonDocument> has been removed from the EntityRepository class. The corresponding initialization in the constructor has also been eliminated. This field was previously used to hold a collection of BsonDocument objects but is no longer needed.

* Remove unused MongoDB.Bson directive from EntityRepository.cs

The `using MongoDB.Bson;` directive was removed from the
`EntityRepository.cs` file as it is no longer needed. The rest
of the file remains unchanged, including the usage of the
`Humanizer` library and the definition of the
`EntityRepository<TEntity>` class within the
`FluentCMS.Repositories.MongoDB` namespace.

* Optimize queries and handle null results in EntityRepo

Added an index on the `Id` field in the `EntityRepository` constructor to ensure efficient querying. Modified the `GetAll` and `GetByIds` methods to handle null results from `ToListAsync` by returning an empty list instead of using the spread operator.