#2074 resolve warnings of admin plugins (#2076)
* #2074 resolve warnings of admin plugins

* Refactor MULTI_SELECT handling in ModelHelpers

Refactor the handling of the MULTI_SELECT case in the ModelHelpers class.
The original single-line return statement is replaced with a more readable
multi-line format. Additionally, ensure each element in the
valuesDict[fieldModel.Name] array is converted to a string, with a fallback
to an empty string if the element is null.

* Refactor PageCreatePlugin properties and null handling

Removed `Urls`, `LayoutOptions`, and `PageOptions` properties from `PageCreatePlugin` class. Decorated `OpenNewPage` with `[SupplyParameterFromQuery(Name = "openNewPage")]` attribute. Replaced null-coalescing operator (`??`) with null-forgiving operator (`!`) for `Title` and `Path` properties.

* Simplify object initializers and use null-forgiving operator

Updated `PageListPlugin.razor.cs` to simplify the `NavigateTo`
method call by using a simplified object initializer for the `Id`
parameter. Replaced null-coalescing with the null-forgiving
operator in list construction. Modified `PageUpdatePlugin.razor.cs`
to use the null-forgiving operator for `SelectOption` object
creation, asserting non-nullability of properties.

---------

Co-authored-by: Amir Pournasserian <amir.pournasserian@live.com>