Remove GetHierarchyBySiteId method and related API clients
The `GetHierarchyBySiteId` method in `PageController.cs` has been removed. This method was responsible for fetching a hierarchical list of pages by site ID.

Corresponding changes have been made in `GeneratedApiClients.cs` to remove the `GetHierarchyBySiteIdAsync` method and its overloads from the `FluentCMS.Web.ApiClients` namespace. This includes the removal of the method's implementation, which handled HTTP GET requests to the `api/Page/GetHierarchyBySiteId` endpoint.

Additionally, the `PageDetailResponseListIApiResult` class has been removed from `GeneratedApiClients.cs`. This class was used to represent the response structure for the `GetHierarchyBySiteId` method.

These changes indicate that the functionality to retrieve a hierarchical list of pages by site ID is no longer supported or needed in the application.